move stores to a single folder
